refactor: file hierarchy
- components should all be in components folder
- entry point should not always return a promise
- entry point should be primary component
- remove old RisingYouth terminology

BREAKING CHANGE
