+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@takingitglobal/a11ize",
   "repository": "https://github.com/takingitglobal/a11ize.git",
-  "main": "index.js",
+  "main": "src/index.js",
   "jsdelivr": "dist/with-react-cjs/index.js",
   "version": "0.0.0-development",
   "description": "A React and JS Accessibility Toolkit",
